Code um SoftwareRegister ueberarbeitet und Tests fuer RegisterAllocator erstellt.

+++ src/Register/RegisterActivityLogger.h
@@ -28,7 +28,7 @@ public:
 
 	void logRegisterModification( Type newValue )
 	{
-		logFile << newValue << std::endl;
+		logFile << (int)newValue << std::endl;
 	}
 
 	virtual ~RegisterActivityLogger()
